First upload.
The metrics used in every context were specific for the case study done on Burundi's public hospitals. They are changed depending on the kind of study you want to use this tool. 
The tool was developed using HTML, CSS, PHP, JAVASCRIPT, SQL.
Remember to download a database management system (XAMPP) to use the database.